quotient computed by piop
