+++ w3f-plonk-common/src/piop.rs
@@ -1,4 +1,4 @@
-use ark_ff::PrimeField;
+use ark_ff::{FftField, PrimeField};
 use ark_poly::univariate::DensePolynomial;
 use ark_poly::Evaluations;
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
@@ -9,6 +9,8 @@ use crate::domain::{Domain, EvaluatedDomain};
 use crate::{ColumnsCommited, ColumnsEvaluated};
 
 pub trait ProverPiop<F: PrimeField, C: Commitment<F>> {
+    const N_CONSTRAINTS: usize;
+
     type Commitments: ColumnsCommited<F, C>;
     type Evaluations: ColumnsEvaluated<F>;
     type Instance: CanonicalSerialize + CanonicalDeserialize;
@@ -29,6 +31,16 @@ pub trait ProverPiop<F: PrimeField, C: Commitment<F>> {
     // Constraint polynomials in evaluation form.
     fn constraints(&self) -> Vec<Evaluations<F>>;
 
+    fn quotient(&self, alphas: &[F]) -> DensePolynomial<F> {
+        let constraint_polys = self.constraints();
+        // Aggregate constraint polynomials in evaluation form...
+        let agg_constraint_poly = aggregate_evaluations(&constraint_polys, &alphas);
+        // ...and then interpolate (to save some FFTs).
+        let agg_constraint_poly = agg_constraint_poly.interpolate();
+        let quotient_poly = self.domain().divide_by_vanishing_poly(&agg_constraint_poly);
+        quotient_poly
+    }
+
     // 'Linearized' parts of constraint polynomials.
     // For a constraint of the form C = C(c1(X),...,ck(X),c1(wX),...,ck(wX)), where ci's are of degree n,
     // and an evaluation point z, it is a degree n polynomial r = C(c1(z),...,ck(z),c1(X),...,ck(X)).
@@ -39,6 +51,16 @@ pub trait ProverPiop<F: PrimeField, C: Commitment<F>> {
 
     // The result of the computation.
     fn result(&self) -> Self::Instance;
+}
+
+fn aggregate_evaluations<F: FftField>(polys: &[Evaluations<F>], coeffs: &[F]) -> Evaluations<F> {
+    assert_eq!(coeffs.len(), polys.len());
+    polys
+        .iter()
+        .zip(coeffs.iter())
+        .map(|(p, &c)| p * c)
+        .reduce(|acc, p| &acc + &p)
+        .unwrap()
 }
 
 pub trait VerifierPiop<F: PrimeField, C: Commitment<F>> {
